Add drivetrain velocity control, implement Xbox controller indexer
This commit closes #13 (xbox controller button indexer) and addresses #10 (drivetrain velocity control).

Added velocity control for Talon controlled drivetrain. PID still needs to be tuned and the maximum velocity values updated in the code, and additionally a "gear shift" has not yet been implemented. Additionally, a manual override button has been added to the dashboard, which defaults the motors to the known working WPI Arcade Drive mode.

Implement Xbox controller button indexer. This is a purely aesthetic change, which improves the clarity of the code because buttons on controllers are accessed via name instead of an obscure number.