Add drive to distance closed loop VI, add auton mode options
This commit addresses issue #2, and needs testing and tuning.

Added a VI called in Auton.vi which will drive forward a specified number of feet. The function uses motion magic to handle closed loop control, and should therefore be pretty accurate. NOTE: this function is not tested, and some of the units may need to be updated/converted.

Cleaned up Auton.vi and added 3 test loops for testing functionality of both drive by time and drive by distance.